Update CopyFileUtils.(h/cpp) 1) Refactor

+++ headers/copy_file_test/CopyFileUtils.h
@@ -1,14 +1,12 @@
 #ifndef OS_HEADERS_COPY_FILE_TEST_COPYFILEUTILS_H_
 #define OS_HEADERS_COPY_FILE_TEST_COPYFILEUTILS_H_
 
-#include <cstdio>
 #include <string>
-#include <stdexcept>
 
 #include "../cmd/CmdUtils.h"
 
 namespace copy_file_utils {
-bool copyFile(int bufferSize, const std::string& inputFileName, const std::string& outputFileName);
+bool copyFile(size_t bufferSize, const std::string& inputFileName, const std::string& outputFileName);
 }
 
 #endif
